Split the settings menu into components for more readability
